Refactor task-related UI components and backend APIs
This update removes `TaskSidebar` and related legacy code while introducing modularized components for task settings, assignees, and labels. Adds `LabelsPopup` and `AssigneesPopup` for managing labels and assignees efficiently. Refactors backend with new API endpoints for adding/removing labels to tasks and enhances task update handling across the Kanban board.
