+++ assets/components/custom/kanban/tasks/modal/task-assignee-section.tsx
@@ -1,20 +1,21 @@
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar.tsx";
-import { Badge } from "@/components/ui/badge.tsx";
-import { Button } from "@/components/ui/button.tsx";
-import { Command, CommandEmpty, CommandGroup, CommandInput, CommandItem } from "@/components/ui/command.tsx";
-import { FormField, FormItem, FormLabel } from "@/components/ui/form.tsx";
-import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover.tsx";
-import type { TaskFormData } from "@/hooks/useTaskForm.ts";
-import type { Team } from "@/types.ts";
-import { Plus, X } from "lucide-react";
-import type { UseFormReturn } from "react-hook-form";
+import {Avatar, AvatarFallback, AvatarImage} from "@/components/ui/avatar.tsx";
+import {Badge} from "@/components/ui/badge.tsx";
+import {Button} from "@/components/ui/button.tsx";
+import {FormField, FormItem, FormLabel} from "@/components/ui/form.tsx";
+import type {TaskFormData} from "@/hooks/useTaskForm.ts";
+import type {Team} from "@/types.ts";
+import {Plus, X} from "lucide-react";
+import type {UseFormReturn} from "react-hook-form";
+import {useState} from "react";
+import {AssigneesPopup} from "@/components/custom/kanban/tasks/assignees-popup.tsx";
 
 interface TaskAssigneeSectionProps {
   form: UseFormReturn<TaskFormData>;
   teamMembers: Team["members"];
 }
 
-export function TaskAssigneeSection({ form, teamMembers }: TaskAssigneeSectionProps) {
+export function TaskAssigneeSection({form, teamMembers}: TaskAssigneeSectionProps) {
+  const [isModalOpen, setIsModalOpen] = useState(false);
   const watchedAssigneeIds = form.watch("assigneeIds") || [];
 
   function addAssignee(userId: number) {
@@ -51,7 +52,7 @@ export function TaskAssigneeSection({ form, teamMembers }: TaskAssigneeSectionPr
                     className="flex items-center gap-2 py-1 px-2"
                   >
                     <Avatar className="w-4 h-4">
-                      <AvatarImage src="" alt={user.name} />
+                      <AvatarImage src="" alt={user.name}/>
                       <AvatarFallback className="text-xs">
                         {user.name
                           .split(" ")
@@ -67,7 +68,7 @@ export function TaskAssigneeSection({ form, teamMembers }: TaskAssigneeSectionPr
                       className="h-3 w-3 p-0 hover:bg-destructive hover:text-destructive-foreground"
                       onClick={() => removeAssignee(userId)}
                     >
-                      <X className="w-2 h-2" />
+                      <X className="w-2 h-2"/>
                     </Button>
                   </Badge>
                 );
@@ -75,49 +76,24 @@ export function TaskAssigneeSection({ form, teamMembers }: TaskAssigneeSectionPr
             </div>
 
             {/* Add Assignee Popover */}
-            <Popover>
-              <PopoverTrigger asChild>
-                <Button
-                  type="button"
-                  variant="outline"
-                  size="sm"
-                  className="h-8"
-                >
-                  <Plus className="w-3 h-3 mr-1" />
-                  Add Assignee
-                </Button>
-              </PopoverTrigger>
-              <PopoverContent className="w-64 p-0">
-                <Command>
-                  <CommandInput placeholder="Search team members..." />
-                  <CommandEmpty>No team members found.</CommandEmpty>
-                  <CommandGroup>
-                    {teamMembers
-                      .filter((user) => !watchedAssigneeIds.includes(user.id))
-                      .map((user) => (
-                        <CommandItem
-                          key={user.id}
-                          onSelect={() => addAssignee(user.id)}
-                          className="cursor-pointer"
-                        >
-                          <div className="flex items-center gap-2">
-                            <Avatar className="w-6 h-6">
-                              <AvatarImage src="" alt={user.name} />
-                              <AvatarFallback className="text-xs">
-                                {user.name
-                                  .split(" ")
-                                  .map((n) => n[0])
-                                  .join("")}
-                              </AvatarFallback>
-                            </Avatar>
-                            <span>{user.name}</span>
-                          </div>
-                        </CommandItem>
-                      ))}
-                  </CommandGroup>
-                </Command>
-              </PopoverContent>
-            </Popover>
+            <AssigneesPopup
+              open={isModalOpen}
+              onOpenChange={setIsModalOpen}
+              teamMembers={teamMembers}
+              onAssigneeAdd={(user) => addAssignee(user.id)}
+              excludedIds={watchedAssigneeIds}
+              onClose={() => setIsModalOpen(false)}
+            >
+              <Button
+                type="button"
+                variant="outline"
+                size="sm"
+                className="h-8"
+              >
+                <Plus className="w-3 h-3 mr-1"/>
+                Add Assignee
+              </Button>
+            </AssigneesPopup>
           </div>
         </FormItem>
       )}
